+++ spec/ig_markets/cli/commands/prices_command_spec.rb
@@ -27,7 +27,7 @@ describe IGMarkets::CLI::Main do
   it 'lists recent prices' do
     historical_price_result = build :historical_price_result
 
-    expect(market).to receive(:historical_prices).with(resolution: :day, max: 1).and_return(historical_price_result)
+    expect(market).to receive(:historical_prices).with(resolution: :day, number: 1).and_return(historical_price_result)
 
     expect { cli(epic: 'A', resolution: :day, number: 1).prices }.to output(<<-END
 +-------------------------+------+-------+-----+------+
@@ -49,14 +49,14 @@ END
 
     options = {
       resolution: :day,
-      from: Time.new(2014, 1, 2, 3, 4, 0, '+00:00'),
-      to: Time.new(2014, 2, 3, 4, 5, 0, '+00:00')
+      from: Time.new(2014, 1, 2, 3, 4, 5, '+00:00'),
+      to: Time.new(2014, 2, 3, 4, 5, 6, '+00:00')
     }
 
     expect(market).to receive(:historical_prices).with(options).and_return(historical_price_result)
 
     expect do
-      cli(epic: 'A', resolution: :day, start_date: '2014-01-02T03:04+00:00', end_date: '2014-02-03T04:05+00:00').prices
+      cli(epic: 'A', resolution: :day, from: '2014-01-02T03:04:05+00:00', to: '2014-02-03T04:05:06+00:00').prices
     end.to output(<<-END
 +-------------------------+------+-------+-----+------+
 |                    Prices for A                     |